[feat] bot 요청 응답 스키마 편집 #22
