+++ backend/schemas/request/bot.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel, Field
-from typing import List, datetime
+from typing import List, Optional, EmailStr
 
 
 class BotDataItem(BaseModel):
@@ -7,14 +7,23 @@ class BotDataItem(BaseModel):
     filename: str = Field(..., example="faq.pdf")
     type: int = Field(..., example=0, description="파일 타입 (예: 0: json, 1: pdf)")
     storage_url: str = Field(..., example="https://storage.example.com/faq.pdf")
-    
-class BotDetailResponse(BaseModel):
-    bot_id: str = Field(..., example="bot_001")
+
+class BotCreateRequest(BaseModel):
     company_name: str = Field(..., example="GenBot")
     usage: str = Field(..., example="문의")
     first_text: str = Field(..., example="안녕하세요! GenBot의 문의봇입니다.")
-    email: str = Field(..., example="user@example.com")
+    email: EmailStr = Field(..., example="user@example.com")
     cs_number: str = Field(..., example="1522-0000")
     data: List[BotDataItem]
-    created_at: datetime = Field(..., example="2025-07-14T12:00:00Z")
-    updated_at: datetime = Field(..., example="2025-07-14T12:10:00Z")
+
+class BotUpdateRequest(BaseModel):
+    company_name: Optional[str] = Field(None, example="GenBot1")
+    usage: Optional[str] = Field(None, example="업무 문의")
+    first_text: Optional[str] = Field(None, example="안녕하세요. 무엇을 도와드릴까요?")
+    email: Optional[EmailStr] = Field(None, example="update@example.com")
+    cs_number: Optional[str] = Field(None, example="1544-0000")
+    data: Optional[List[BotDataItem]]
+
+# 임시
+class DataValidationRequest(BaseModel):
+    file_names: List[str] = Field(..., example=["faq.pdf", "test.json"])
